Parsing/IR: late conversion of the subdenomination value to improve too many digit detector (fix #297)
